refactor: moving the graph client
